RavenDB-21381 - Addressed PR Comments
